Move stategraph nodes to separate files
