+++ llm/nodes/quality_control.py
@@ -1,0 +1,216 @@
+import json
+import logging
+
+from llm.LLMDefinition import LLM
+from llm.nodes.node_logger import node_logger
+from llm.tools.Tools_aggregator import get_tools
+
+logger = logging.getLogger("quality_control_node")
+logger.setLevel(logging.INFO)
+
+
+# --- Quality Control Node (QC) ---
+
+
+@node_logger(
+    "quality_control",
+    input_keys=["user_query", "out_of_scope_result", "keywords"],
+    output_keys=[
+        "qc_decision",
+        "qc_tool_result",
+        "keywords",
+        "has_filter_instructions",
+    ],
+)
+def quality_control_node(state):
+    """
+    Perform quality control and filter detection on the user query.
+    Args:
+        state (dict): The current agent state.
+    Returns:
+        dict: Updated state with QC decision, tool result, keywords, and filter instructions flag.
+    """
+    tools = get_tools()
+    tool_map = {getattr(tool, "name", None): tool for tool in tools}
+    qc_decision = "accept"  # Default
+    qc_tool_result = None
+    try:
+        result = state.get("out_of_scope_result")
+        user_query = state.get("user_query", "")
+        keywords = state.get("keywords", [])
+
+        # Extract keywords from out_of_scope_result if available
+        if result:
+            try:
+                parsed = json.loads(result)
+                if isinstance(parsed, dict):
+                    if parsed.get("status") == "out_of_scope":
+                        qc_decision = "out_of_scope"
+                        qc_tool_result = result
+                        state["qc_decision"] = qc_decision
+                        state["qc_tool_result"] = qc_tool_result
+                        # Don't return here - let the LLM QC decision process continue
+                    elif parsed.get("status") == "valid" and "keywords" in parsed:
+                        keywords = parsed["keywords"]
+                        state["keywords"] = keywords
+            except Exception as e:
+                logger.error(f"Error parsing out_of_scope_result: {e}")
+
+        # Update keywords in state
+        state["keywords"] = keywords
+
+        # Use LLM to intelligently detect filter instructions
+        filter_detection_prompt = f"""
+        You are an academic research assistant. Analyze the user query to determine if it contains filter instructions.
+
+        Filter instructions include:
+        - Date/time constraints: "after 2020", "before 2018", "published since 2022", "between 2019-2023"
+        - Citation constraints: "highly cited", "more than 50 citations", "well-cited papers"
+        - Author constraints: "by author X", "from researcher Y", "authored by"
+        - Journal/conference constraints: "published in Nature", "from conference X", "in journal Y"
+        - Impact constraints: "high impact", "top journals", "prestigious venues"
+        - Similarity constraints: "highly relevant", "closely related", "similar to"
+        - Specific metrics: "fwci > 5", "citation percentile > 90", "impact factor > 10"
+
+        User query: "{user_query}"
+
+        Respond with ONLY a JSON object:
+        {{
+        "has_filter_instructions": true/false,
+        "reason": "brief explanation of why"
+        }}
+        """
+
+        try:
+            filter_response = LLM.invoke(filter_detection_prompt)
+            filter_response_content = (
+                filter_response.content
+                if hasattr(filter_response, "content")
+                else str(filter_response)
+            )
+            filter_result = (
+                json.loads(filter_response_content)
+                if isinstance(filter_response_content, str)
+                else filter_response_content
+            )
+
+            # Handle potential list response
+            if isinstance(filter_result, list) and len(filter_result) > 0:
+                filter_result = filter_result[0]
+
+            has_filter_instructions = False
+            reason = "No reason provided"
+            if isinstance(filter_result, dict):
+                has_filter_instructions = filter_result.get(
+                    "has_filter_instructions", False
+                )
+                reason = filter_result.get("reason", "No reason provided")
+
+            state["has_filter_instructions"] = has_filter_instructions
+            logger.info(f"Filter detection: {has_filter_instructions} - {reason}")
+
+        except Exception as e:
+            logger.error(f"Error in filter detection: {e}")
+            state["has_filter_instructions"] = False
+
+        # LLM-driven QC decision
+        qc_prompt = f"""
+        You are an academic research assistant. Given the following user query and keywords, decide which action to take:
+        - If the query is valid and specific, respond with 'accept'.
+        - If the query is vague, respond with 'reformulate'.
+        - If the query is too broad, respond with 'narrow'.
+        - If the query is too narrow, respond with 'broaden'.
+        - If the query contains multiple topics, respond with 'split'.
+        - If the query is not a valid research topic, respond with 'out_of_scope'.
+
+        User query: "{user_query}"
+        Keywords: {keywords}
+
+        Respond with a JSON object:
+        {{
+        "qc_decision": "accept" | "reformulate" | "broaden" | "narrow" | "split" | "out_of_scope",
+        "reason": "..."
+        }}
+        """
+        qc_response = LLM.invoke(qc_prompt)
+        qc_response_content = (
+            qc_response.content if hasattr(qc_response, "content") else str(qc_response)
+        )
+        qc_result = (
+            json.loads(qc_response_content)
+            if isinstance(qc_response_content, str)
+            else qc_response_content
+        )
+        # If qc_result is a list, use the first element if possible
+        if isinstance(qc_result, list) and len(qc_result) > 0:
+            qc_result = qc_result[0]
+        if not isinstance(qc_result, dict):
+            qc_result = {}
+        qc_decision = qc_result.get("qc_decision", "accept")
+        state["qc_decision"] = qc_decision
+        state["qc_decision_reason"] = qc_result.get("reason", "")
+        # Call the appropriate tool if needed
+        if qc_decision == "reformulate":
+            tool = tool_map.get("reformulate_query")
+            if tool:
+                qc_tool_result = tool.invoke(
+                    {"query_description": user_query, "keywords": keywords}
+                )
+                try:
+                    tool_result = (
+                        json.loads(qc_tool_result)
+                        if isinstance(qc_tool_result, str)
+                        else qc_tool_result
+                    )
+                    if (
+                        "result" in tool_result
+                        and "refined_keywords" in tool_result["result"]
+                    ):
+                        state["keywords"] = tool_result["result"]["refined_keywords"]
+                except Exception as e:
+                    logger.error(f"Error in ...: {e}")
+        elif qc_decision == "split":
+            tool = tool_map.get("multi_step_reasoning")
+            if tool:
+                qc_tool_result = tool.invoke({"query_description": user_query})
+        elif qc_decision == "broaden":
+            tool = tool_map.get("retry_broaden")
+            if tool:
+                qc_tool_result = tool.invoke(
+                    {"query_description": user_query, "keywords": keywords}
+                )
+                try:
+                    tool_result = (
+                        json.loads(qc_tool_result)
+                        if isinstance(qc_tool_result, str)
+                        else qc_tool_result
+                    )
+                    if "broadened_keywords" in tool_result:
+                        state["keywords"] = tool_result["broadened_keywords"]
+                except Exception as e:
+                    logger.error(f"Error in ...: {e}")
+        elif qc_decision == "narrow":
+            tool = tool_map.get("narrow_query")
+            if tool:
+                qc_tool_result = tool.invoke(
+                    {"query_description": user_query, "keywords": keywords}
+                )
+                try:
+                    tool_result = (
+                        json.loads(qc_tool_result)
+                        if isinstance(qc_tool_result, str)
+                        else qc_tool_result
+                    )
+                    if "narrowed_keywords" in tool_result:
+                        state["keywords"] = tool_result["narrowed_keywords"]
+                except Exception as e:
+                    logger.error(f"Error in ...: {e}")
+        elif qc_decision == "accept":
+            tool = tool_map.get("accept")
+            if tool:
+                qc_tool_result = tool.invoke({"confirmation": "yes"})
+        state["qc_tool_result"] = qc_tool_result
+    except Exception as e:
+        state["error"] = f"QC node error: {e}"
+        qc_decision = "error"
+    return state
